# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AUIB Academic Advisor API",
    "description": "Degree planning for AUIB students. Planning endpoints are stateless and store nothing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/rules": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Rules to review, each beside its source sentence",
        "operationId": "list_rules_api_v1_admin_rules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "show",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "needs_review",
                "partial",
                "unparsed",
                "overridden",
                "source_changed",
                "reviewed"
              ],
              "type": "string",
              "default": "needs_review",
              "title": "Show"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuleListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rules/check": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Check a rule written in the rule language without saving it",
        "operationId": "check_rule_api_v1_admin_rules_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/rules/{rule_id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Correct a rule by hand (F0.3); the correction survives re-imports",
        "operationId": "correct_rule_api_v1_admin_rules__rule_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin: catalog"
        ],
        "summary": "Delete a rule an admin added",
        "operationId": "delete_rule_api_v1_admin_rules__rule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rules/{rule_id}/approve": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Mark the rule in force as checked",
        "operationId": "approve_rule_api_v1_admin_rules__rule_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/rules/{rule_id}/override": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Drop a correction and go back to the parsed rule",
        "operationId": "remove_override_api_v1_admin_rules__rule_id__override_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/imports": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Program imports, newest first",
        "operationId": "imports_api_v1_admin_imports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportRunOut"
                  },
                  "title": "Response Imports Api V1 Admin Imports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin actions, newest first (F9.3)",
        "operationId": "audit_api_v1_admin_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditOut"
                  },
                  "title": "Response Audit Api V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminRuleListOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AdminRuleOut"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "total",
          "counts",
          "rules"
        ],
        "title": "AdminRuleListOut"
      },
      "AdminRuleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "course_title": {
            "type": "string",
            "title": "Course Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "unparsed_text": {
            "type": "string",
            "title": "Unparsed Text"
          },
          "parsed_rule": {
            "type": "string",
            "title": "Parsed Rule"
          },
          "parsed_english": {
            "type": "string",
            "title": "Parsed English"
          },
          "override_rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Rule"
          },
          "override_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Note"
          },
          "effective_rule": {
            "type": "string",
            "title": "Effective Rule"
          },
          "effective_english": {
            "type": "string",
            "title": "Effective English"
          },
          "reviewed": {
            "type": "boolean",
            "title": "Reviewed"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "source_changed": {
            "type": "boolean",
            "title": "Source Changed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_code",
          "course_title",
          "kind",
          "source_text",
          "status",
          "unparsed_text",
          "parsed_rule",
          "parsed_english",
          "override_rule",
          "override_note",
          "effective_rule",
          "effective_english",
          "reviewed",
          "reviewed_by",
          "reviewed_at",
          "source_changed"
        ],
        "title": "AdminRuleOut"
      },
...
      "AuditOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "at",
          "actor",
          "action",
          "target",
          "detail"
        ],
        "title": "AuditOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "program_id": {
            "type": "string",
            "title": "Program Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published": {
            "type": "boolean",
            "title": "Published"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          }
        },
        "type": "object",
        "required": [
          "id",
          "program_id",
          "started_at",
          "status",
          "published",
          "actor",
          "report"
        ],
        "title": "ImportRunOut"
      },
...
      "RuleCheckIn": {
        "properties": {
          "rule": {
            "type": "string",
            "maxLength": 2000,
            "title": "Rule"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule"
        ],
        "title": "RuleCheckIn"
      },
      "RuleCheckOut": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule"
          },
          "english": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "unknown_courses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Courses"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "RuleCheckOut"
      },
...
      "RuleUpdateIn": {
        "properties": {
          "rule": {
            "type": "string",
            "maxLength": 2000,
            "title": "Rule",
            "description": "Rule in the admin rule language, or NONE"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule"
        ],
        "title": "RuleUpdateIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin token (ADVISOR_ADMIN_TOKEN)",
        "scheme": "bearer"
      }
    }
  }
}